Added an email to patients when driver is updated for their appointment.

+++ app/models/appointment.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class Appointment
   include Mongoid::Document
   field :datetime, type: String
@@ -15,18 +17,22 @@ class Appointment
 
   before_save :get_est_time
   before_update :get_est_time
+  after_update :email_updates
 
   def self.clean_past_appointments
     Appointment.each do |appt|
-      if DateTime.strptime(appt.datetime, '%Y-%m-%d %H:%M').to_date.past?
-        appt.destroy
-      end
+      appt.destroy if DateTime.strptime(appt.datetime, '%Y-%m-%d %H:%M').to_date.past?
     end
   end
 
   def get_est_time
-    url="https://maps.googleapis.com/maps/api/directions/json?origin=" + self.location[0].coordinates[0].to_s + "," + self.location[0].coordinates[1].to_s + "&destination=" + self.location[1].coordinates[0].to_s + "," + self.location[1].coordinates[1].to_s + "&key=" + Rails.application.credentials.gmap_geocode_api_kep
+    url = 'https://maps.googleapis.com/maps/api/directions/json?origin=' + location[0].coordinates[0].to_s + ',' + location[0].coordinates[1].to_s + '&destination=' + location[1].coordinates[0].to_s + ',' + location[1].coordinates[1].to_s + '&key=' + Rails.application.credentials.gmap_geocode_api_kep
     response = HTTParty.get(url).parsed_response
-    self.est_time = ((response['routes'].first['legs'].first['duration']['value'])/60).round + 15 # BUFFER TIME
+    self.est_time = ((response['routes'].first['legs'].first['duration']['value']) / 60).round + 15 # BUFFER TIME
+  end
+
+  def email_updates
+    UserMailer.with(appt: appt).ride_updated_email.deliver
+    UserMailer.with(appt: appt).ride_assigned_email.deliver
   end
 end
